Fix bug

+++ ros2bag/ros2bag/verb/record.py
@@ -132,7 +132,7 @@ def add_recorder_arguments(parser: ArgumentParser) -> None:
         '--qos-profile-overrides-path', type=FileType('r'),
         help='Path to a yaml file defining overrides of the QoS profile for specific topics.')
     parser.add_argument(
-        '--timeout-for-delay', type=int, default=90,
+        '--timeout-for-delay', type=float, default=90,
         help='Timeout to stop auto-discovery. By default %(default)d seconds.')
 
     # Core config
